--- src/components/breadcrumb/index.tsx
@@ -1,1 +1,0 @@
-export { Breadcrumb } from "./breadcrumb";
