+++ src/components/breadcrumbs/index.tsx
@@ -1,0 +1,1 @@
+export { Breadcrumbs } from "./breadcrumbs";
